+++ condense/src/main/java/com/condense/annotation/CommandFilter.java
@@ -3,6 +3,7 @@ package com.condense.annotation;
 import jakarta.inject.Qualifier;
 
 import java.lang.annotation.ElementType;
+import java.lang.annotation.Repeatable;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
@@ -29,6 +30,7 @@ import java.lang.annotation.Target;
 @Qualifier
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE, ElementType.FIELD, ElementType.PARAMETER, ElementType.METHOD})
+@Repeatable(CommandFilters.class)
 public @interface CommandFilter {
     String value();
 }